+++ packages/checkbox/src/checkbox.tsx
@@ -7,9 +7,9 @@ import {
   SystemStyleObject,
   ThemingProps,
   useMultiStyleConfig,
-  WithChakraProps,
+  HTMLChakraProps,
 } from "@chakra-ui/system"
-import { callAll, cx, Omit, __DEV__ } from "@chakra-ui/utils"
+import { callAll, cx, Omit, __DEV__, dataAttr } from "@chakra-ui/utils"
 import * as React from "react"
 import { useCheckboxGroupContext } from "./checkbox-group"
 import { CheckboxIcon } from "./checkbox-icon"
@@ -47,7 +47,7 @@ type Omitted =
   | "onBlur"
   | "value"
 
-type StyledControlProps = Omit<WithChakraProps<"div">, Omitted>
+type StyledControlProps = Omit<HTMLChakraProps<"div">, Omitted>
 
 type BaseInputProps = Pick<
   PropsOf<"input">,
@@ -158,7 +158,7 @@ export const Checkbox = forwardRef<CheckboxProps, "input">(function Checkbox(
   return (
     <StyledContainer
       __css={styles.container}
-      data-disabled={state.isDisabled}
+      data-disabled={dataAttr(state.isDisabled)}
       className={_className}
       {...htmlProps}
     >