+++ packages/hooks/src/use-focus-on-show.ts
@@ -1,21 +1,26 @@
-import { focus, FocusableElement, getAllFocusable } from "@chakra-ui/utils"
+import {
+  contains,
+  focus,
+  FocusableElement,
+  getAllFocusable,
+} from "@chakra-ui/utils"
 import * as React from "react"
 import { useEventListener } from "./use-event-listener"
-import { useLatestRef } from "./use-latest-ref"
+import { useUpdateEffect } from "./use-update-effect"
 
-export interface UseConditionalFocusOptions {
+export interface UseFocusOnShowOptions {
   visible?: boolean
   shouldFocus?: boolean
   preventScroll?: boolean
   focusRef?: React.RefObject<FocusableElement>
 }
 
-const defaultOptions: UseConditionalFocusOptions = {
+const defaultOptions: UseFocusOnShowOptions = {
   preventScroll: true,
   shouldFocus: false,
 }
 
-export function useConditionalFocus<T extends HTMLElement>(
+export function useFocusOnShow<T extends HTMLElement>(
   target: React.RefObject<T> | T,
   options = defaultOptions,
 ) {
@@ -24,14 +29,10 @@ export function useConditionalFocus<T extends HTMLElement>(
 
   const autoFocus = shouldFocus && visible
 
-  const onFocus = () => {
-    if (
-      !element ||
-      !autoFocus ||
-      element.contains(document.activeElement) ||
-      element === document.activeElement
-    )
-      return
+  const onFocus = React.useCallback(() => {
+    if (!element || !autoFocus) return
+
+    if (contains(element, document.activeElement as HTMLElement)) return
 
     if (focusRef?.current) {
       focus(focusRef.current, { preventScroll })
@@ -41,13 +42,11 @@ export function useConditionalFocus<T extends HTMLElement>(
         focus(tabbableEls[0], { preventScroll })
       }
     }
-  }
+  }, [autoFocus, preventScroll])
 
-  const onFocusRef = useLatestRef(onFocus)
-
-  React.useEffect(() => {
-    onFocusRef.current()
-  }, [])
+  useUpdateEffect(() => {
+    onFocus()
+  }, [onFocus])
 
   useEventListener("transitionend", onFocus, element)
 }