+++ packages/icon/src/icon.tsx
@@ -1,8 +1,8 @@
 import {
   chakra,
+  ChakraProps,
   forwardRef,
   SystemStyleObject,
-  WithChakraProps,
 } from "@chakra-ui/system"
 import { cx, __DEV__ } from "@chakra-ui/utils"
 import * as React from "react"
@@ -26,7 +26,9 @@ const fallbackIcon = {
   viewBox: "0 0 24 24",
 }
 
-export interface IconProps extends WithChakraProps<"svg"> {}
+export interface IconProps
+  extends Omit<React.SVGAttributes<SVGElement>, keyof ChakraProps>,
+    ChakraProps {}
 
 export const Icon = forwardRef<IconProps, "svg">(function Icon(props, ref) {
   const {