+++ packages/input/src/input-element.tsx
@@ -3,12 +3,12 @@ import {
   forwardRef,
   SystemStyleObject,
   useStyles,
-  WithChakraProps,
+  HTMLChakraProps,
 } from "@chakra-ui/system"
 import { cx, __DEV__ } from "@chakra-ui/utils"
 import * as React from "react"
 
-export interface InputElementProps extends WithChakraProps<"div"> {
+export interface InputElementProps extends HTMLChakraProps<"div"> {
   placement?: "left" | "right"
 }