+++ packages/layout/src/center.tsx
@@ -1,7 +1,7 @@
-import { chakra, WithChakraProps } from "@chakra-ui/system"
+import { chakra, HTMLChakraProps } from "@chakra-ui/system"
 import { __DEV__ } from "@chakra-ui/utils"
 
-export interface CenterProps extends WithChakraProps<"div"> {}
+export interface CenterProps extends HTMLChakraProps<"div"> {}
 
 /**
  * React component used to horizontally and vertically center its child.