+++ packages/layout/src/code.tsx
@@ -4,12 +4,12 @@ import {
   omitThemingProps,
   ThemingProps,
   useStyleConfig,
-  WithChakraProps,
+  HTMLChakraProps,
 } from "@chakra-ui/system"
 import { cx, __DEV__ } from "@chakra-ui/utils"
 import * as React from "react"
 
-export interface CodeProps extends WithChakraProps<"code">, ThemingProps {}
+export interface CodeProps extends HTMLChakraProps<"code">, ThemingProps {}
 
 /**
  * React component to render inline code snippets.