+++ packages/menu/src/use-menu.ts
@@ -3,7 +3,6 @@ import { useDescendant, useDescendants } from "@chakra-ui/descendant"
 import {
   useControllableState,
   useDisclosure,
-  useEventListener,
   useFocusOnHide,
   useId,
   useIds,
@@ -25,8 +24,8 @@ import {
   getValidChildren,
   isArray,
   isString,
-  merge,
   mergeRefs,
+  mergeWith,
   normalizeEventKey,
   removeItem,
 } from "@chakra-ui/utils"
@@ -166,7 +165,11 @@ export function useMenu(props: UseMenuProps) {
     }
   }, [isOpen])
 
-  useFocusOnHide(menuRef, { focusRef: buttonRef, visible: isOpen })
+  useFocusOnHide(menuRef, {
+    focusRef: buttonRef,
+    visible: isOpen,
+    shouldFocus: true,
+  })
 
   /**
    * Generate unique ids for menu's list and button
@@ -191,32 +194,18 @@ export function useMenu(props: UseMenuProps) {
   const refocus = () => {
     const hasFocusWithin = menuRef.current?.contains(document.activeElement)
     const shouldRefocus = isOpen && !hasFocusWithin
+
     if (!shouldRefocus) return
-    requestAnimationFrame(() => {
-      const el = domContext.descendants[focusedIndex]?.element
-      el?.focus({ preventScroll: true })
-    })
-  }
 
-  useEventListener("transitionend", refocus, menuRef.current)
-
-  const onTransitionEnd = () => {
-    /**
-     * If we use CSS for transitioning this component, there would be no
-     * need to dispatch a custom event. This is only useful for JS only
-     * motion libraries like `framer-motion` to `react-spring`.
-     *
-     * They usually provide an `onRest` or `onAnimationComplete` callback we can
-     * use to trigger the custom `transitionend` event.
-     */
-    menuRef.current?.dispatchEvent(new Event("transitionend"))
+    const el = domContext.descendants[focusedIndex]?.element
+    el?.focus({ preventScroll: true })
   }
 
   return {
     openAndFocusMenu,
     openAndFocusFirstItem,
     openAndFocusLastItem,
-    onTransitionEnd,
+    onTransitionEnd: refocus,
     domContext,
     popper,
     buttonId,
@@ -352,7 +341,7 @@ export function useMenuList(
 
 export function useMenuPositioner(props: any = {}) {
   const { popper, isOpen } = useMenuContext()
-  return merge(popper.getPopperProps(props), {
+  return mergeWith(popper.getPopperProps(props), {
     style: { visibility: isOpen ? "visible" : "hidden" },
   })
 }
@@ -457,6 +446,7 @@ export function useMenuItem(
     closeOnSelect,
     onClose,
     menuRef,
+    isOpen,
   } = menu
 
   const ref = useRef<HTMLDivElement>(null)
@@ -506,12 +496,13 @@ export function useMenuItem(
   const trulyDisabled = isDisabled && !isFocusable
 
   useUpdateEffect(() => {
+    if (!isOpen) return
     if (isFocused && !trulyDisabled && ref.current) {
       focus(ref.current)
     } else if (document.activeElement !== menuRef.current) {
       menuRef.current?.focus()
     }
-  }, [isFocused, trulyDisabled, menuRef])
+  }, [isFocused, trulyDisabled, menuRef, isOpen])
 
   const tabbable = useClickable({
     onClick,
@@ -582,7 +573,7 @@ export function useMenuOptionGroup(props: UseMenuOptionGroupProps) {
     type = "radio",
     value: valueProp,
     defaultValue,
-    onChange,
+    onChange: onChangeProp,
     ...htmlProps
   } = props
 
@@ -593,10 +584,10 @@ export function useMenuOptionGroup(props: UseMenuOptionGroupProps) {
   const [value, setValue] = useControllableState({
     defaultValue: defaultValue ?? fallback,
     value: valueProp,
-    onChange,
+    onChange: onChangeProp,
   })
 
-  const handleChange = useCallback(
+  const onChange = useCallback(
     (selectedValue: string) => {
       if (type === "radio" && isString(value)) {
         setValue(selectedValue)
@@ -626,7 +617,7 @@ export function useMenuOptionGroup(props: UseMenuOptionGroupProps) {
     if ((child.type as any).id !== "MenuItemOption") return child
 
     const onClick = (event: MouseEvent) => {
-      handleChange(child.props.value)
+      onChange(child.props.value)
       child.props.onClick?.(event)
     }