+++ packages/switch/src/switch.tsx
@@ -6,7 +6,7 @@ import {
   SystemStyleObject,
   ThemingProps,
   useMultiStyleConfig,
-  WithChakraProps,
+  HTMLChakraProps,
 } from "@chakra-ui/system"
 import { cx, dataAttr, __DEV__ } from "@chakra-ui/utils"
 import * as React from "react"
@@ -15,7 +15,7 @@ type Omitted = "defaultChecked" | "checked" | "onChange"
 
 export interface SwitchProps
   extends Omit<UseCheckboxProps, "isIndeterminate">,
-    Omit<WithChakraProps<"label">, Omitted>,
+    Omit<HTMLChakraProps<"label">, Omitted>,
     ThemingProps {}
 
 export const Switch = forwardRef<SwitchProps, "input">((props, ref) => {