+++ packages/tag/src/tag.tsx
@@ -8,12 +8,12 @@ import {
   ThemingProps,
   useMultiStyleConfig,
   useStyles,
-  WithChakraProps,
+  HTMLChakraProps,
 } from "@chakra-ui/system"
 import { __DEV__ } from "@chakra-ui/utils"
 import * as React from "react"
 
-export interface TagProps extends WithChakraProps<"span">, ThemingProps {}
+export interface TagProps extends HTMLChakraProps<"span">, ThemingProps {}
 
 /**
  * The tag component is used to label or categorize UI elements.
@@ -43,7 +43,7 @@ if (__DEV__) {
   Tag.displayName = "Tag"
 }
 
-export interface TagLabelProps extends WithChakraProps<"span"> {}
+export interface TagLabelProps extends HTMLChakraProps<"span"> {}
 
 export const TagLabel = forwardRef<TagLabelProps, "span">(function TagLabel(
   props,
@@ -93,7 +93,7 @@ if (__DEV__) {
 }
 
 export interface TagCloseButtonProps
-  extends Omit<WithChakraProps<"button">, "disabled"> {
+  extends Omit<HTMLChakraProps<"button">, "disabled"> {
   isDisabled?: boolean
 }