+++ packages/transition/src/scale-fade.tsx
@@ -1,4 +1,4 @@
-import { cx, merge, __DEV__ } from "@chakra-ui/utils"
+import { cx, mergeWith, __DEV__ } from "@chakra-ui/utils"
 import { AnimatePresence, HTMLMotionProps, motion } from "framer-motion"
 import * as React from "react"
 import { EASINGS, MotionVariants } from "./__utils"
@@ -66,7 +66,7 @@ export const ScaleFade = React.forwardRef<HTMLDivElement, ScaleFadeProps>(
     const show = unmountOnExit ? isOpen && unmountOnExit : true
     const custom = { initialScale, reverse }
 
-    const motionProps = merge(scaleFadeConfig, {
+    const motionProps = mergeWith(scaleFadeConfig, {
       custom,
       animate: isOpen || unmountOnExit ? "enter" : "exit",
     })