+++ packages/hooks/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 1.0.0-rc.7 (2020-10-25)
+
+**Note:** Version bump only for package @chakra-ui/hooks
+
+
+
+
+
+# 1.0.0-rc.6 (2020-10-25)
+
+**Note:** Version bump only for package @chakra-ui/hooks
+
+
+
+
+
 # 1.0.0-rc.5 (2020-09-27)
 
 **Note:** Version bump only for package @chakra-ui/hooks